🐞 FIX: 카트에서 사이즈나 수량을 수정할 때에 카트데이터가 삭제되는 버그 수정

+++ src/main/java/com/supernova/fashionnova/domain/cart/CartService.java
@@ -140,12 +140,15 @@ public class CartService {
         if (existingCartOptional.isPresent()) {
             // 이미 존재하면 수량 증가
             Cart existingCart = existingCartOptional.get();
-            existingCart.incrementCount(cart.getCount());
-            existingCart.incrementTotalPrice(cart.getTotalPrice());
-            cartRepository.save(existingCart);
 
-            // 기존 장바구니 항목 삭제
-            cartRepository.delete(cart);
+            if (!cart.equals(existingCart)) {
+                existingCart.incrementCount(cart.getCount());
+                existingCart.incrementTotalPrice(cart.getTotalPrice());
+
+                cartRepository.delete(cart);
+            }else{
+                existingCart.updateCountPrice(dto.getCount());
+            }
         } else {
             // 존재하지 않으면 새로 추가
             cart.setProductDetail(newProductDetail);
